pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiir"
version = "1.1.0"
description = "AI-powered Incident Response analysis toolset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1,<9.0",
    "openai>=1.0,<2.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0,<2.0",
    "rich>=13.0,<14.0",
    "keyring>=25.0,<26.0",
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "jinja2>=3.1,<4.0",
    "python-multipart>=0.0.9,<1.0",
    "aiofiles>=23.0,<25.0",
    "json-repair>=0.58.6,<1.0",
]

[project.scripts]
aiir = "aiir.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aiir"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

